% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{coacross}
\alias{coacross}
\title{Coalesce across columns.}
\usage{
coacross(...)
}
\description{
Additional function to coalesce all the columns with the phosphopeptides
to remove all the columns with NAs.
}
