% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{phospho_positions}
\alias{phospho_positions}
\title{Find phosphosites positions.}
\usage{
phospho_positions(
  df,
  ptm = c("(CARBAMIDOMETHYL)", "(TMT6PLEX)", "(OXIDATION)")
)
}
\arguments{
\item{df}{Dataframe with the phosphoproteomics data.}

\item{ptm}{Vector with the name of the postranlational modifications in the peptides.
different from phosphorylations that are going to be removed.}
}
\value{
A dataframe with the positions of the phosphosites in the peptide.
}
\description{
This function received a data frame with the phosphoproteomics data and the
sequence of the modified peptide in a column named 'peptide' to remove all
the undesired modifications that are not a phosphorylation, and identify the
position of the phosphorylation site in the phosphopeptide. The function is
intended to detect a maximum of two phosphosites per peptide.
The input dataframe must have a column call "Pho" wich indicates if the
peptide has a phosphorylation in it or not.
}
\examples{
data(KupfCells)
phospho_positions(KupfCells)

}
