% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoFind.R
\name{PhosphoFind}
\alias{PhosphoFind}
\title{Find phosphosites}
\usage{
PhosphoFind(df, psp_db)
}
\arguments{
\item{df}{A phosphoproteomics dataframe with the columns previously especified}

\item{psp_db}{Dataframe with the PhosphoSitePlus phosphorylation data for the
organism}
}
\value{
A list with the first element as a dataframe with the phosphoproteomics
data with the identified phosphorylation sites. The second element is a vector
with the protein Uniprot IDs that have not been identified.
}
\description{
This function receives a phosphoproteomics data dataframe with, at least,
the following columns:
"Pho": Column indicating if the peptides contains a phosphorylation
"protein": Uniprot (ACC ID) identifier of the protein
"Phospho_sequence": sequence of the peptide
"Phosphosite_1": position in the peptide of the first phosphosite
"Phosphosite_2": position in the peptide of the second phosphosite (if any)
}
\examples{
data(PhosphoHeart)
psp_db <- load_PSP_db(organism = "mouse")
# Obtain list with phosphorylation sites and unidentified proteins:
Phospholist <- PhosphoFind(df = PhosphoHeart, psp_db = psp_db)
# Extract dataframe with the identified phosphorylation sites:
Phosphosites <- Phospholist[[1]]
# Extract names of proteins not identified:
No_ID <- Phospholist[[2]]

}
